game-bot/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
mss>=9.0.0
pyautogui>=0.9.54
pyyaml>=6.0
pytesseract>=0.3.10
keyboard>=0.13.5
